ai-service/requirements.txt
flask
flask-cors
ultralytics
opencv-python
mediapipe
numpy
Pillow
